perf: remove tooltip on select and move it to option
