+++ packages/components/src/components/option/index.ts
@@ -1,4 +1,5 @@
 import '../icon';
+import '../tooltip';
 import Option from './option.component';
 import { TAG_NAME } from './option.constants';
