+++ packages/components/src/components/select/index.ts
@@ -1,7 +1,6 @@
 import '../icon';
 import '../popover';
 import '../text';
-import '../tooltip';
 import Select from './select.component';
 import { TAG_NAME } from './select.constants';
